simple sdk sample
ignore log files

Add local test

update
