+++ .gitignore
@@ -3,3 +3,5 @@ target/
 local.log
 .idea
 *.iml
+.DS_Store
+**/logs/*
