+++ android/testng-examples/src/test/java/com/browserstack/run_first_test/AppiumTest.java
@@ -1,0 +1,28 @@
+package com.browserstack.run_first_test;
+
+import java.net.URL;
+
+import org.openqa.selenium.remote.DesiredCapabilities;
+import org.testng.annotations.AfterMethod;
+import org.testng.annotations.BeforeMethod;
+
+import io.appium.java_client.AppiumDriver;
+import io.appium.java_client.android.AndroidDriver;
+import io.appium.java_client.android.AndroidElement;
+
+
+public class AppiumTest {
+
+    public AppiumDriver<AndroidElement> driver;
+
+    @BeforeMethod(alwaysRun=true)
+    public void setUp() throws Exception {
+        DesiredCapabilities capabilities = new DesiredCapabilities();
+        driver = new AndroidDriver(new URL("http://127.0.0.1:4723/wd/hub"),capabilities);
+    }
+
+    @AfterMethod(alwaysRun=true)
+    public void tearDown() throws Exception {
+        driver.quit();
+    }
+}
