+++ android/testng-examples/src/test/java/com/browserstack/run_first_test/FirstTest.java
@@ -1,16 +1,16 @@
 package com.browserstack.run_first_test;
 
+import java.util.List;
+
+import org.openqa.selenium.support.ui.ExpectedConditions;
+import org.openqa.selenium.support.ui.WebDriverWait;
+import org.testng.Assert;
+import org.testng.annotations.Test;
+
 import io.appium.java_client.MobileBy;
 import io.appium.java_client.android.AndroidElement;
 
-import org.openqa.selenium.support.ui.WebDriverWait;
-import org.openqa.selenium.support.ui.ExpectedConditions;
-
-import org.testng.Assert;
-import org.testng.annotations.Test;
-import java.util.List;
-
-public class FirstTest extends BrowserStackTestNGTest {
+public class FirstTest extends AppiumTest {
 
     @Test
     public void test() throws Exception {
