--- android/testng-examples/src/test/resources/com/browserstack/run_local_test/local.conf.json
@@ -1,21 +1,0 @@
-{
-  "server": "hub-cloud.browserstack.com",
-  "username": "BROWSERSTACK_USERNAME",
-  "access_key": "BROWSERSTACK_ACCESS_KEY",
-
-  "capabilities": {
-    "project": "First TestNg Android Project",
-    "build": "browserstack-build-1",
-    "name": "local_test",
-    "browserstack.debug": true,
-    "browserstack.local": true,
-    "app": "bs://<app-id>"
-  },
-
-  "environments": [
-    { 
-      "device": "Google Pixel 3",
-      "os_version": "9.0"
-    }
-  ]
-}
