--- android/testng-examples/src/test/resources/com/browserstack/run_local_test/local.testng.xml
@@ -1,9 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE suite SYSTEM "http://testng.org/testng-1.0.dtd">
-<suite name="Local">
-    <test name="LocalTest">
-        <classes>
-            <class name="com.browserstack.run_local_test.LocalTest"/>
-        </classes>
-    </test>
-</suite>
