--- android/testng-examples/src/test/resources/com/browserstack/run_parallel_test/parallel.conf.json
@@ -1,24 +1,0 @@
-{
-  "server": "hub-cloud.browserstack.com",
-  "username": "BROWSERSTACK_USERNAME",
-  "access_key": "BROWSERSTACK_ACCESS_KEY",
-
-  "capabilities": {
-    "project": "First TestNg Android Project",
-    "build": "browserstack-build-1",
-    "name": "parallel_test",
-    "browserstack.debug": true,
-    "app": "bs://<app-id>"
-  },
-
-  "environments": [
-    {
-      "device": "Google Pixel 3",
-      "os_version": "9.0"
-    },
-    {
-      "device": "Samsung Galaxy S10e",
-      "os_version": "9.0"
-    }
-  ]
-}
