+++ ios/testng-examples/src/test/java/com/browserstack/run_first_test/AppiumTest.java
@@ -1,0 +1,27 @@
+package com.browserstack.run_first_test;
+
+import java.net.URL;
+
+import org.openqa.selenium.remote.DesiredCapabilities;
+import org.testng.annotations.AfterMethod;
+import org.testng.annotations.BeforeMethod;
+
+import io.appium.java_client.ios.IOSDriver;
+import io.appium.java_client.ios.IOSElement;
+
+
+public class AppiumTest {
+
+    public IOSDriver<IOSElement> driver;
+
+    @BeforeMethod(alwaysRun=true)
+    public void setUp() throws Exception {
+        DesiredCapabilities capabilities = new DesiredCapabilities();
+        driver = new IOSDriver(new URL("http://127.0.0.1:4723/wd/hub"),capabilities);
+    }
+
+    @AfterMethod(alwaysRun=true)
+    public void tearDown() throws Exception {
+        driver.quit();
+    }
+}
