--- ios/testng-examples/src/test/java/com/browserstack/run_parallel_test/ParallelTest.java
@@ -1,31 +1,0 @@
-package com.browserstack.run_parallel_test;
-
-import org.testng.Assert;
-import org.testng.annotations.Test;
-
-import org.openqa.selenium.support.ui.WebDriverWait;
-import org.openqa.selenium.support.ui.ExpectedConditions;
-
-import io.appium.java_client.MobileBy;
-import io.appium.java_client.ios.IOSElement;
-
-
-public class ParallelTest extends BrowserStackTestNGTest {
-
-  @Test
-  public void test() throws Exception {
-    IOSElement textButton = (IOSElement) new WebDriverWait(driver, 30).until(
-        ExpectedConditions.elementToBeClickable(MobileBy.AccessibilityId("Text Button")));
-    textButton.click();
-    IOSElement textInput = (IOSElement) new WebDriverWait(driver, 30).until(
-        ExpectedConditions.elementToBeClickable(MobileBy.AccessibilityId("Text Input")));
-    textInput.sendKeys("hello@browserstack.com"+"\n");
-
-    Thread.sleep(5000);
-
-    IOSElement textOutput = (IOSElement) new WebDriverWait(driver, 30).until(
-        ExpectedConditions.elementToBeClickable(MobileBy.AccessibilityId("Text Output")));
-
-    Assert.assertEquals(textOutput.getText(),"hello@browserstack.com");
-  }
-}
