--- ios/testng-examples/src/test/resources/com/browserstack/run_local_test/local.conf.json
@@ -1,21 +1,0 @@
-{
-  "server": "hub-cloud.browserstack.com",
-  "username": "BROWSERSTACK_USERNAME",
-  "access_key": "BROWSERSTACK_ACCESS_KEY",
-
-  "capabilities": {
-    "project": "First TestNg iOS Project",
-    "build": "browserstack-build-1",
-    "name": "local_test",
-    "browserstack.debug": true,
-    "browserstack.local": true,
-    "app": "bs://<app-id>"
-  },
-
-  "environments": [
-    { 
-      "device": "iPhone 11 Pro",
-      "os_version": "13"
-    }
-  ]
-}
