--- ios/testng-examples/src/test/resources/com/browserstack/run_parallel_test/parallel.conf.json
@@ -1,24 +1,0 @@
-{
-  "server": "hub-cloud.browserstack.com",
-  "username": "BROWSERSTACK_USERNAME",
-  "access_key": "BROWSERSTACK_ACCESS_KEY",
-
-  "capabilities": {
-    "project": "First TestNg iOS Project",
-    "build": "browserstack-build-1",
-    "name": "parallel_test",
-    "browserstack.debug": true,
-    "app": "bs://<app-id>"
-  },
-
-  "environments": [
-    {
-      "device": "iPhone 11 Pro",
-      "os_version": "13"
-    },
-    {
-      "device": "iPhone 11 Pro Max",
-      "os_version": "13"
-    }
-  ]
-}
